cookbooks/python/browser-use/requirements.txt
swarmkit
python-dotenv
rich
